General Feedback Cleanup
Rename Samples to Examples
Add a bunch of documentation about prefabs and scripts.
Update sub-module to most recent build.
Make SignIn button show a "Not Enabled" message when you're missing an XboxServices.config file.
Add better error handling to UnityTaskExtensions stuff.

+++ Assets/Xbox Live/Scripts/Stats/DoubleStat.cs
@@ -12,7 +12,7 @@ using Microsoft.Xbox.Services.Stats.Manager;
 [Serializable]
 public class DoubleStat : StatBase<double>
 {
-    public void Multiply(double multiplier)
+    public void Multiply(float multiplier)
     {
         this.SetValue(Value * multiplier);
     }
